fix(utils): include directory getters in module.exports
Modified the module.exports object in utils.js to include the directory getters directly as getter properties. This addresses an issue where the getters were defined on the exports object using Object.defineProperty but were not included in the final module.exports object.

Previously, when other modules imported utils.js and tried to access the directory getters (like utils.API_CONTRACTS_DIR), they would receive undefined even though the branch was set correctly. This was because the getters were not properly exported.

This change aims to make the branch-aware directory getters accessible to other modules by including them directly in the module.exports object. The getters maintain the same functionality but are now properly exported.

This is a supplementary fix to the branch-aware alignment mutations, addressing an export issue that prevented the getters from being accessible to other modules.